Moved classes that inherited SubnetInfo
IP4Subnet and IP6Subnet were moved to the until package.
